feat(server): allow running server without a PSK
Adds functionality to the public protocol to accept requests without a PSK if running in a mode without PSKs.

Also modifies the server CLI to be more succinct when displaying the connection command.
Only the required flags are displayed now, so PSK flags will not be shown if they are not set.